register form works fine
