+++ xumage/templates/login.html
@@ -1,3 +1,6 @@
+{% load static %}
+{% csrf_token %}
+
 <!DOCTYPE html>
 <!-- saved from url=(0062)login.html -->
 <html lang="ru">
@@ -7,11 +10,11 @@
 
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Три Вкуса - Меню</title>
-	<link rel="icon" href="/img/logotip.svg" type="image/x-icon">
+	<link rel="icon" href="{% static '/img/logotip.svg" type="image/x-icon">
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@9/swiper-bundle.min.css" />
-	<link rel="stylesheet" href="css/style.css" />
-	<link rel="stylesheet" href="css/reset.css" />
-	<link rel="stylesheet" href="css/adaptive.css" />
+	<link rel="stylesheet" href="{% static 'css/style.css' %}" />
+	<link rel="stylesheet" href="{% static 'css/reset.css' %}" />
+	<link rel="stylesheet" href="{% static 'css/adaptive.css' %}" />
 </head>
 
 <body>
@@ -20,25 +23,25 @@
 
 		<header class="header">
 			<div class="container header__container">
-				<a href="index.html" class="logo header__logo">
-					<img src="img/logotip.svg" alt="Логотип" class="header__logo_img-link" />
+				<a href="/" class="logo header__logo">
+					<img src="{% static 'img/logotip.svg' %}" alt="Логотип" class="header__logo_img-link" />
 				</a>
 				<ul class="header__nav">
 					<li class="header__nav_item">
-						<a href="index.html">Меню</a>
+						<a href="/">Меню</a>
 					</li>
 					<li class="header__nav_item">
-						<a href="sale.html">Акции</a>
+						<a href="/sales">Акции</a>
 					</li>
 					<li class="header__nav_item">
-						<a href="contacts.html">Контакты</a>
+						<a href="/contacts">Контакты</a>
 					</li>
 					<li class="header__nav_item">
-						<a href="delivery.html">Доставка</a>
+						<a href="/delivery">Доставка</a>
 					</li>
 					<li class="header__nav_item">
 						<a href="tel: +79177777777">
-							<img src="img/telephone.svg" alt="+7 (917) 777-77-77" class="header__nav_item-img" />
+							<img src="{% static 'img/telephone.svg' %}" alt="+7 (917) 777-77-77" class="header__nav_item-img" />
 							+7 (917) 777-77-77
 						</a>
 					</li>
@@ -47,7 +50,7 @@
 					<div class="burger__line"></div>
 				</button>
 				<div class="login">
-					<a class="login-link" href="login.html">
+					<a class="login-link" href="/login">
 						<svg class="login-link__icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 64" aria-labelledby="title"
 							aria-describedby="desc" role="img" xmlns:xlink="http://www.w3.org/1999/xlink" width="48" height="48">
 							<title>Login</title>
@@ -67,34 +70,34 @@
 			<nav class="mobile-menu__nav">
 				<ul class="mobile-menu__nav-list">
 					<li class="mobile-menu__nav-list_item">
-						<a href="index.html">Меню</a>
+						<a href="/">Меню</a>
 					</li>
 					<li class="mobile-menu__nav-list_item">
-						<a href="sale.html">Акции</a>
+						<a href="/sales">Акции</a>
 					</li>
 					<li class="mobile-menu__nav-list_item">
-						<a href="contacts.html">Контакты</a>
+						<a href="/contacts">Контакты</a>
 					</li>
 					<li class="mobile-menu__nav-list_item">
-						<a href="delivery.html">Доставка</a>
+						<a href="/delivery">Доставка</a>
 					</li>
 				</ul>
 			</nav>
 			<div class="mobile-menu__contacts">
 				<div class="mobile-menu__contacts_item">
 					<a href="#">
-						<img src="img/telephone.svg" alt="tel" class="mobile-menu__contacts_item_icon">
+						<img src="{% static 'img/telephone.svg' %}" alt="tel" class="mobile-menu__contacts_item_icon">
 						+7 (917) 777-77-77
 					</a>
 				</div>
 				<div class="mobile-menu__contacts_item">
 					<a href="#">
-						<img src="img/vk-button.png" alt="vk" class="mobile-menu__contacts_item_icon">
+						<img src="{% static 'img/vk-button.png' %}" alt="vk" class="mobile-menu__contacts_item_icon">
 						Вконтакте
 					</a>
 				</div>
 				<div class="mobile-menu__contacts_item">
-					<img src="img/location.svg" alt="location" class="mobile-menu__contacts_item_icon">
+					<img src="{% static 'img/location.svg' %}" alt="location" class="mobile-menu__contacts_item_icon">
 					Б. Матросова 27, Салават
 				</div>
 			</div>
@@ -117,7 +120,7 @@
 							</div>
 							<label class="data-input">
 								<input type="text">
-								<img class="data-edit-img" src="/img/data-edit.svg" alt="Изменить данные"></label>
+								<img class="data-edit-img" src="{% static '/img/data-edit.svg' %}" alt="Изменить данные"></label>
 						</div>
 						<div class="inpt">
 							<div class="input-text">
@@ -125,7 +128,7 @@
 							</div>
 							<label class="data-input">
 								<input type="email">
-								<img class="data-edit-img" src="/img/data-edit.svg" alt="Изменить данные"></label>
+								<img class="data-edit-img" src="{% static '/img/data-edit.svg' %}" alt="Изменить данные"></label>
 						</div>
 						<div class="inpt">
 							<div class="input-text">
@@ -133,7 +136,7 @@
 							</div>
 							<label class="data-input">
 								<input type="tel">
-								<img class="data-edit-img" src="/img/data-edit.svg" alt="Изменить данные">
+								<img class="data-edit-img" src="{% static '/img/data-edit.svg' %}" alt="Изменить данные">
 							</label>
 						</div>
 
@@ -155,29 +158,29 @@
 		<footer class="footer">
 			<div class="container footer-top__container">
 				<ul class="footer__info">
-					<li class="footer__info_item"><img class="footer__info_item-img" src="/img/clock.svg" alt=""> Круглосуточно
+					<li class="footer__info_item"><img class="footer__info_item-img" src="{% static '/img/clock.svg' %}" alt=""> Круглосуточно
 					</li>
-					<li class="footer__info_item"><img class="footer__info_item-img" src="/img/geo.svg" alt=""> Б. Матросова 27,
+					<li class="footer__info_item"><img class="footer__info_item-img" src="{% static '/img/geo.svg' %}" alt=""> Б. Матросова 27,
 						Салават</li>
 					<li class="footer__info_item"><a href="tel:+79178004477"><img class="footer__info_item-img"
-								src="/img/telephone.svg" alt=""> +7 (917) 777-77-77</a></li>
+								src="{% static '/img/telephone.svg' %}" alt=""> +7 (917) 777-77-77</a></li>
 				</ul>
 			</div>
 
 			<div class="container footer-bottom__container">
 				<ul class="footer__list">
-					<li class="footer__list_link"><a href="/about.html">О нас</a></li>
+					<li class="footer__list_link"><a href="/about">О нас</a></li>
 					<li class="footer__list_link"><a href="#">Публичная оферта</a></li>
 					<li class="footer__list_link"><a href="#">Политика конфиденциальности</a></li>
 					<li class="footer__list_link"><a target="_blank" href="https://vk.com/trivkusaslv"><img
-								class="footer__list_link-img" src="/img/vk-button.png" alt="ВКонтакте"></a></li>
+								class="footer__list_link-img" src="{% static '/img/vk-button.png' %}" alt="ВКонтакте"></a></li>
 				</ul>
 			</div>
 		</footer>
 
 	</div>
 
-	<script src="js/script.js"></script>
+	<script src="{% static 'js/script.js' %}"></script>
 </body>
 
 </html>
