+++ xumage/treevkusa/views.py
@@ -1,41 +1,16 @@
 from django.shortcuts import render
-from django.http import HttpResponse
-from treevkusa.models import *
+from django.http import HttpResponse, HttpResponseRedirect
+from .crud import *
 
-def get_some_data_from_db(food_type: str) -> dict:
-    response_json_object: dict = { food_type + 's': [] }
-
-    for food in FoodMenu.objects.all().filter(type=food_type):
-        response_json_object[food_type + 's'].append(
-            {
-                'name': food.name,
-                'desc': food.description,
-                'weight': food.weight,
-                'price': food.price,
-                
-            }
-        )
-    print(response_json_object)
-    return response_json_object
-
-def get_all_foods() -> dict:
-    response_json_object: dict = {}
-
-    for food_type in ['burger', 'hotdog', 'pizza', 'shaurma', 'drink']:
-        response_json_object[food_type + 's'] = []
-        for food in FoodMenu.objects.all().filter(type=food_type):
-            response_json_object[food_type + 's'].append(
-                {
-                    'name': food.name,
-                    'desc': food.description,
-                    'weight': food.weight,
-                    'price': food.price,
-                    'img_path': food.image_path
-                }
-            )
-    print(response_json_object)
-    return response_json_object
-
+def postuser(request):
+    name = request.POST.get('name')
+    email = request.POST.get('email')
+    phone_number = request.POST.get('tel')
+    password = request.POST.get('password')
+    verify_password = request.POST.get('verify_password')
+    add_user(name=name, password=password, email=email, phone_number=phone_number)
+    del name,email,phone_number,password,verify_password
+    return HttpResponseRedirect("/login")
 
 def index_page(request):
     return render(request, 'index.html', get_all_foods())
